reorganize schema for necc group
