+++ ce_nome_s/__init__.py
@@ -61,8 +61,6 @@ from baseclasses.chemical_energy import (
     PhaseFluorometryOxygen,
     PumpRateMeasurement,
     LinearSweepVoltammetry,
-    CENECCElectrode,
-    PotentiometryGasChromatographyMeasurement
 )
 
 from baseclasses.helper.utilities import create_archive, rewrite_json, find_sample_by_id
@@ -444,60 +442,8 @@ class CE_NOME_DocumentationTool(DocumentationTool, EntryData):
                 envs.to_excel(writer, sheet_name='environments', index=False)
                 setups.to_excel(writer, sheet_name='setups', index=False)
 
-class CE_NECC_Electrode(CENECCElectrode, EntryData):
-    m_def = Section(
-        a_eln=dict(
-            hide=["users", "origin", "elemental_composition",
-                  "components", "chemical_composition_or_formulas"],
-            properties=dict(
-                order=[
-                    "name",
-                    "lab_id",
-                    "recipe_id",
-                    "recipe_type",
-                    "element",
-                    "deposition_method"
-                ])),
-        label_quantity='sample_id')
 
 # %%####################################### Measurements
-
-class CE_NECC_PotentiometryGasChromatographyMeasurement(PotentiometryGasChromatographyMeasurement, EntryData):
-    m_def = Section(
-        a_eln=dict(
-            hide=[
-                'lab_id', 'location', 'steps', 'samples', 'atmosphere', 'instruments'
-                ],
-            properties=dict(
-                order=[
-                    'name', 'properties', 'gaschromatographies',
-                    'potentiometry', 'thermocouple', 'results'
-                    ])),
-                a_plot=[{
-                    'label': 'Potential-dependent Faradaic efficiencies',
-                    'x': 'potentiometry/working_electrode_potential',
-                    'y': ['results/gas_results/faradaic_efficiency',
-                          'results/total_flow_rate'],
-                    'layout': {"showlegend": True,
-                               'yaxis': {
-                                   "fixedrange": False},
-                               'xaxis': {
-                                   "fixedrange": False}}},
-                    {
-                        'label': 'Current–voltage characteristics and electrode temperatures',
-                        'x': 'potentiometry/working_electrode_potential',
-                        'y': ['results/gas_results/current',
-                              'thermocouple/temperature_cathode'],
-                        'layout': {"showlegend": True,
-                                   'yaxis': {
-                                       "fixedrange": False},
-                                   'xaxis': {
-                                       "fixedrange": False}},
-                    }])
-
-    def normalize(self, archive, logger):
-        super(CE_NECC_PotentiometryGasChromatographyMeasurement, self).normalize(archive, logger)
-
 
 class Bessy2_KMC2_XASFluorescence(XASFluorescence, EntryData):
     m_def = Section(
